# server/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  dto.BackendDTO:
    properties:
      servers:
        description: 服务器列表，至少需要一个服务器
        items:
          $ref: '#/definitions/dto.ServerDTO'
        minItems: 1
        type: array
    required:
    - servers
    type: object
...
  dto.CertificateDTO:
    properties:
      certName:
        description: 证书名称
        example: my-cert
        type: string
      expireDate:
        description: 过期时间
        type: string
      fingerPrint:
        description: 证书指纹
        type: string
      issuerName:
        description: 颁发机构
        example: Let's Encrypt
        type: string
      privateKey:
        description: 私钥内容
        type: string
      publicKey:
        description: 公钥内容
        type: string
    required:
    - certName
    - expireDate
    - fingerPrint
    - issuerName
    - privateKey
    - publicKey
    type: object
...
  dto.CreateSiteRequest:
    description: 创建站点的请求参数
    properties:
      activeStatus:
        description: 站点状态
        example: true
        type: boolean
      backend:
        allOf:
        - $ref: '#/definitions/dto.BackendDTO'
        description: 后端服务器配置
      certificate:
        allOf:
        - $ref: '#/definitions/dto.CertificateDTO'
        description: 证书信息
      domain:
        description: 域名
        example: example.com
        type: string
      enableHTTPS:
        description: 是否启用HTTPS
        example: false
        type: boolean
      listenPort:
        description: 监听端口
        example: 8080
        maximum: 65535
        minimum: 1
        type: integer
      name:
        description: 站点名称
        example: my-site
        type: string
      wafEnabled:
        description: 是否启用WAF
        example: false
        type: boolean
      wafMode:
        description: WAF模式
        enum:
        - protection
        - observation
        example: observation
        type: string
    required:
    - backend
    - domain
    - listenPort
    - name
    type: object
...
  dto.RunnerControlRequest:
    properties:
      action:
        description: 控制动作
        enum:
        - start
        - stop
        - restart
        - force_stop
        - reload
        type: string
    required:
    - action
    type: object
  dto.RunnerControlResponse:
    properties:
      action:
        description: 执行的动作
        example: start
        type: string
      message:
        description: 操作消息
        example: 运行器已成功启动
        type: string
      state:
        description: 操作后的状态
        example: running
        type: string
      success:
        description: 操作是否成功
        example: true
        type: boolean
    type: object
  dto.RunnerStatusResponse:
    properties:
      isRunning:
        description: 是否正在运行
        example: true
        type: boolean
      state:
        description: 状态：running, stopped, error
        example: running
        type: string
    type: object
  dto.ServerDTO:
    properties:
      host:
        description: 主机地址
        example: backend.example.com
        type: string
      isSSL:
        description: 是否启用SSL
        example: false
        type: boolean
      port:
        description: 端口
        example: 80
        maximum: 65535
        minimum: 1
        type: integer
    required:
    - host
    - port
    type: object
  dto.SiteListResponse:
    description: 站点列表响应
    properties:
      items:
        description: 站点列表
        items:
          $ref: '#/definitions/model.Site'
        type: array
      total:
        description: 总数
        type: integer
    type: object
  dto.SiteResponse:
    description: 站点信息响应
    properties:
      activeStatus:
        description: 站点是否激活
        type: boolean
      backend:
        allOf:
        - $ref: '#/definitions/model.Backend'
        description: 后端服务器配置
      certificate:
        allOf:
        - $ref: '#/definitions/model.Certificate'
        description: 证书信息
      createdAt:
        type: string
      domain:
        description: 域名，如 a.com
        type: string
      enableHTTPS:
        description: 是否启用HTTPS
        type: boolean
      id:
        description: 站点ID
        type: string
      listenPort:
        description: 监听端口，如 9000
        type: integer
      name:
        description: 站点名称
        type: string
      updatedAt:
        type: string
      wafEnabled:
        description: 是否启用WAF
        type: boolean
      wafMode:
        allOf:
        - $ref: '#/definitions/model.WAFMode'
        description: WAF防护模式
    type: object
  dto.UpdateSiteRequest:
    description: 更新站点的请求参数
    properties:
      activeStatus:
        description: 站点状态
        example: true
        type: boolean
      backend:
        allOf:
        - $ref: '#/definitions/dto.BackendDTO'
        description: 后端服务器配置
      certificate:
        allOf:
        - $ref: '#/definitions/dto.CertificateDTO'
        description: 证书信息
      domain:
        description: 域名
        example: example.com
        type: string
      enableHTTPS:
        description: 是否启用HTTPS
        example: false
        type: boolean
      listenPort:
        description: 监听端口
        example: 8080
        maximum: 65535
        minimum: 1
        type: integer
      name:
        description: 站点名称
        example: my-site
        type: string
      wafEnabled:
        description: 是否启用WAF
        example: false
        type: boolean
      wafMode:
        description: WAF模式
        enum:
        - protection
        - observation
        example: observation
        type: string
    type: object
...
  model.Backend:
    properties:
      servers:
        description: 服务器列表
        items:
          $ref: '#/definitions/model.Server'
        type: array
    type: object
  model.Certificate:
    properties:
      certName:
        description: 证书名称/别名
        type: string
      expireDate:
        description: 证书过期日期
        type: string
      fingerPrint:
        description: 证书指纹
        type: string
      issuerName:
        description: 颁发机构
        type: string
      privateKey:
        description: 私钥内容（PEM格式）
        type: string
      publicKey:
        description: 公钥内容（PEM格式）
        type: string
    type: object
...
  model.ErrResponse:
    description: 错误的API响应标准格式
    properties:
      code:
        example: 400
        type: integer
      error:
        example: 参数错误
        type: string
      message:
        example: 请求参数错误
        type: string
      requestId:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      success:
        example: false
        type: boolean
      timestamp:
        example: "2023-01-01T12:00:00Z"
        type: string
    type: object
  model.ErrResponseDontShowError:
    description: 错误的API响应标准格式,不展示 error
    properties:
      code:
        example: 400
        type: integer
      message:
        example: 请求参数错误
        type: string
      requestId:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      success:
        example: false
        type: boolean
      timestamp:
        example: "2023-01-01T12:00:00Z"
        type: string
    type: object
...
  model.Server:
    properties:
      host:
        description: 主机地址，如 IP 或域名
        type: string
      isSSL:
        description: 是否启用SSL
        type: boolean
      port:
        description: 端口
        type: integer
    type: object
  model.Site:
    properties:
      activeStatus:
        description: 站点是否激活
        type: boolean
      backend:
        allOf:
        - $ref: '#/definitions/model.Backend'
        description: 后端服务器配置
      certificate:
        allOf:
        - $ref: '#/definitions/model.Certificate'
        description: 证书信息
      createdAt:
        type: string
      domain:
        description: 域名，如 a.com
        type: string
      enableHTTPS:
        description: 是否启用HTTPS
        type: boolean
      id:
        description: 站点ID
        type: string
      listenPort:
        description: 监听端口，如 9000
        type: integer
      name:
        description: 站点名称
        type: string
      updatedAt:
        type: string
      wafEnabled:
        description: 是否启用WAF
        type: boolean
      wafMode:
        allOf:
        - $ref: '#/definitions/model.WAFMode'
        description: WAF防护模式
    type: object
  model.SuccessResponse:
    description: 成功的API响应标准格式
    properties:
      code:
        example: 200
        type: integer
      data: {}
      message:
        example: 操作成功
        type: string
      requestId:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      success:
        example: true
        type: boolean
      timestamp:
        example: "2023-01-01T12:00:00Z"
        type: string
    type: object
  model.SuccessResponseNoData:
    description: 成功的API响应标准格式,没有 data
    properties:
      code:
        example: 200
        type: integer
      message:
        example: 操作成功
        type: string
      requestId:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      success:
        example: true
        type: boolean
      timestamp:
        example: "2023-01-01T12:00:00Z"
        type: string
    type: object
...
  model.WAFMode:
    enum:
    - protection
    - observation
    type: string
    x-enum-comments:
      WAFModeObservation: 观察模式
      WAFModeProtection: 防护模式
    x-enum-varnames:
    - WAFModeProtection
    - WAFModeObservation
host: localhost:2333
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: 简单的 Web 应用防火墙管理系统 API
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Simple-WAF API
  version: "1.0"
paths:
  /api/runner/control:
    post:
      consumes:
      - application/json
      description: 执行启动、停止、重启、强制停止或热重载操作
      parameters:
      - description: 运行器控制请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.RunnerControlRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 操作成功
          schema:
            allOf:
            - $ref: '#/definitions/model.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.RunnerControlResponse'
              type: object
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/model.ErrResponse'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/model.ErrResponseDontShowError'
      security:
      - BearerAuth: []
      summary: 控制后台运行器
      tags:
      - 运行器管理
  /api/runner/status:
    get:
      description: 获取WAF后台运行器的运行状态
      produces:
      - application/json
      responses:
        "200":
          description: 获取运行器状态成功
          schema:
            allOf:
            - $ref: '#/definitions/model.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.RunnerStatusResponse'
              type: object
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/model.ErrResponseDontShowError'
      security:
      - BearerAuth: []
      summary: 获取后台运行器状态
      tags:
      - 运行器管理
...
  /api/v1/site:
    get:
      description: 获取所有站点配置列表
      parameters:
      - default: 1
        description: 页码
        in: query
        name: page
        type: integer
      - default: 10
        description: 每页数量
        in: query
        name: size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 获取站点列表成功
          schema:
            allOf:
            - $ref: '#/definitions/model.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.SiteListResponse'
              type: object
        "401":
          description: 未授权访问
          schema:
            $ref: '#/definitions/model.ErrResponseDontShowError'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/model.ErrResponseDontShowError'
      security:
      - BearerAuth: []
      summary: 获取站点列表
      tags:
      - 站点管理
    post:
      consumes:
      - application/json
      description: 创建一个新的站点配置
      parameters:
      - description: 站点信息
        in: body
        name: site
        required: true
        schema:
          $ref: '#/definitions/dto.CreateSiteRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 站点创建成功
          schema:
            allOf:
            - $ref: '#/definitions/model.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.SiteResponse'
              type: object
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/model.ErrResponse'
        "401":
          description: 未授权访问
          schema:
            $ref: '#/definitions/model.ErrResponseDontShowError'
        "403":
          description: 禁止访问
          schema:
            $ref: '#/definitions/model.ErrResponseDontShowError'
        "409":
          description: 域名和端口组合已存在
          schema:
            $ref: '#/definitions/model.ErrResponseDontShowError'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/model.ErrResponseDontShowError'
      security:
      - BearerAuth: []
      summary: 创建新站点
      tags:
      - 站点管理
  /api/v1/site/{id}:
    delete:
      description: 删除指定的站点配置
      parameters:
      - description: 站点ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 站点删除成功
          schema:
            $ref: '#/definitions/model.SuccessResponseNoData'
        "401":
          description: 未授权访问
          schema:
            $ref: '#/definitions/model.ErrResponseDontShowError'
        "403":
          description: 禁止访问
          schema:
            $ref: '#/definitions/model.ErrResponseDontShowError'
        "404":
          description: 站点不存在
          schema:
            $ref: '#/definitions/model.ErrResponseDontShowError'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/model.ErrResponseDontShowError'
      security:
      - BearerAuth: []
      summary: 删除站点
      tags:
      - 站点管理
    get:
      description: 根据ID获取站点详情
      parameters:
      - description: 站点ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 获取站点详情成功
          schema:
            allOf:
            - $ref: '#/definitions/model.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.SiteResponse'
              type: object
        "401":
          description: 未授权访问
          schema:
            $ref: '#/definitions/model.ErrResponseDontShowError'
        "404":
          description: 站点不存在
          schema:
            $ref: '#/definitions/model.ErrResponseDontShowError'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/model.ErrResponseDontShowError'
      security:
      - BearerAuth: []
      summary: 获取单个站点
      tags:
      - 站点管理
    put:
      consumes:
      - application/json
      description: 更新指定站点的配置
      parameters:
      - description: 站点ID
        in: path
        name: id
        required: true
        type: string
      - description: 站点更新信息
        in: body
        name: site
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateSiteRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 站点更新成功
          schema:
            allOf:
            - $ref: '#/definitions/model.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.SiteResponse'
              type: object
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/model.ErrResponse'
        "401":
          description: 未授权访问
          schema:
            $ref: '#/definitions/model.ErrResponseDontShowError'
        "403":
          description: 禁止访问
          schema:
            $ref: '#/definitions/model.ErrResponseDontShowError'
        "404":
          description: 站点不存在
          schema:
            $ref: '#/definitions/model.ErrResponseDontShowError'
        "409":
          description: 域名和端口组合已被其他站点使用
          schema:
            $ref: '#/definitions/model.ErrResponseDontShowError'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/model.ErrResponseDontShowError'
      security:
      - BearerAuth: []
      summary: 更新站点
      tags:
      - 站点管理
...
securityDefinitions:
  BearerAuth:
    description: 使用 Bearer {token} 格式进行身份验证
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"